+++ CHANGELOG.md
@@ -13,9 +13,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 - #2817 - Data Importer failed with per-sort was selected due to attempting to sort an immutable list (introduced in #2772)
 - #2806 - AEM Environment Indicator rendered two times when opening a Experience Fragment variation
-- #2806 - AEM Environment Indicator rendered two times when opening a Experience Fragment variation
 - #2812 - Fixed issue with Reports not reporting accurate Replication status when report is downloaded 
 - #2822 - Resolved OakPal issue 92 - False positive during build
+- #2794 - Added context prefix for redirect rules feature
 
 ### Changed